services/demoapp/pyproject.toml
[project]
name = "demoapp"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "TakiMoysha", email = "takimoysha@users.noreply.github.com" },
]
requires-python = ">=3.12"
dependencies = [
    "litestar-saq>=0.5.2",
    "litestar-users>=1.7.0",
    "litestar-vite>=0.13.1",
    "slugify>=0.0.1",
]
dynamic = ["version"]

[tool.hatch.version]
path = "demoapp/__about__.py"

[project.scripts]
server_demoapp = "demoapp.__main__:run_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
